% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\docType{class}
\name{activity-class}
\alias{activity-class}
\title{An S4 class to represent a Strava activity.}
\description{
An S4 class to represent a Strava activity.
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the Strava activity (character).
By default \code{name = NA_character_}.}

\item{\code{type}}{the type of the Strava activity (character), like "Ride", "Run", etc.
By default \code{type = NA_character_}.}

\item{\code{lat}}{a numeric vector, containing the recorded lattidue coordinates.
By default \code{lat = NULL}.}

\item{\code{lon}}{a numeric vector, containing the recorded longitude coodinates.
By default \code{lon = NULL}.}

\item{\code{distances}}{a numeric vector containing the distances between the lat/lon coodrinates in meter.
Starts at zero.
By default \code{distances = NULL}.}

\item{\code{elevations}}{a numeric vector conaining the elevation in meters, for every recorded point.
By default \code{elevations = NULL}.}

\item{\code{times}}{a numeric vector containing the time differences between recorded points.
Starts at zero.
By default \code{times = NULL}.}
}}

