% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{checkErrors}
\alias{checkErrors}
\title{Check HTTR Response for Errors}
\usage{
checkErrors(response)
}
\arguments{
\item{response}{An httr response object}
}
\description{
\code{checkErrors} raises a warning if there was a problem with the response
status code (i.e. it was not 200)
}

