% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{checkRateLimit}
\alias{checkRateLimit}
\title{Check Rate Limit Status}
\usage{
checkRateLimit(response)
}
\arguments{
\item{response}{An httr response object}
}
\description{
Raises a warning if the user has used more than 90% of short or
long term rate limit
}

