% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{convertCelsiusToFahrenheit}
\alias{convertCelsiusToFahrenheit}
\title{Convert Celsius to Fahrenheit}
\usage{
convertCelsiusToFahrenheit(vect)
}
\arguments{
\item{vect}{Numeric vector in celsius degrees}
}
\description{
Convert Celsius to Fahrenheit
}

