% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{convertMetersPerSecondToMph}
\alias{convertMetersPerSecondToMph}
\title{Convert Meters Per Second to Miles Per Hour}
\usage{
convertMetersPerSecondToMph(vect)
}
\arguments{
\item{vect}{Numeric vector measured in meters per second}
}
\description{
Convert Meters Per Second to Miles Per Hour
}

