% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{convertMetersToMiles}
\alias{convertMetersToMiles}
\title{Convert Meters to Miles}
\usage{
convertMetersToMiles(vect)
}
\arguments{
\item{vect}{Numeric vector measured in meters}
}
\description{
Convert Meters to Miles
}

