% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{convertStravaDate}
\alias{convertStravaDate}
\title{Converts Strava Date-Time Object to POSIXct}
\usage{
convertStravaDate(stravaDateVector)
}
\arguments{
\item{stravaDateVector}{A character vector formatted by Strava}
}
\value{
A Posixct date-time object
}
\description{
Converts Strava Date-Time Object to POSIXct
}

