% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{createOrBind}
\alias{createOrBind}
\title{Either Create Dataframe or Bind, based on row}
\usage{
createOrBind(d, row, out)
}
\arguments{
\item{d}{Integer indicating whether this is the first instance of the data
frame, or a later version}
}
\value{
ou A data frame object
}
\description{
Either Create Dataframe or Bind, based on row
}

