% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_activitytype}
\alias{get_activitytype}
\title{Determine the activity type, based on the Strava gpx file title.}
\usage{
get_activitytype(file)
}
\arguments{
\item{file}{a .gpx file from Strava.}
}
\value{
the activity type, as character.
}
\description{
Determine the activity type, based on the Strava gpx file title.
}
