% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_activity.R
\name{get_countries}
\alias{get_countries}
\title{Determine the country(ies) of the lat/lon coordinates of a Strava activity}
\usage{
get_countries(activity)
}
\arguments{
\item{activity}{an object of S4 class activity.}
}
\value{
character vector of the countries.
}
\description{
Determine the country(ies) of the lat/lon coordinates of a Strava activity
}
