% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_activity.R
\name{get_elevationlevel}
\alias{get_elevationlevel}
\title{Catecorize the slot elevations from an S4 activity object into different levels.}
\usage{
get_elevationlevel(activity, by = 100, levels = 3:30)
}
\arguments{
\item{activity}{an object of S4 class activity.}

\item{by}{numeric value to devide the elevations in meters.
By default \code{by = 100}.}

\item{levels}{numeric vector to determine the corresponding levels.
By default \code{levels = 3:30}.}
}
\value{
the transfromed elevations slot.
}
\description{
Catecorize the slot elevations from an S4 activity object into different levels.
}
