% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_speeds}
\alias{get_speeds}
\title{Calculate the velocities (km/h) of an activity for every measured point.
Using the distance devided by \code{\link{get_timediffs}}.}
\usage{
get_speeds(activity)
}
\arguments{
\item{activity}{an object of S4 class activity.}
}
\value{
the respective velocities.
}
\description{
Calculate the velocities (km/h) of an activity for every measured point.
Using the distance devided by \code{\link{get_timediffs}}.
}
