% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_activity.R
\name{get_timediffs}
\alias{get_timediffs}
\title{Calculate the time difference between measured points in an activity}
\usage{
get_timediffs(activity)
}
\arguments{
\item{activity}{an object of S4 class activity.}
}
\value{
a numeric vecotor of the corresponding time differences.
}
\description{
Calculate the time difference between measured points in an activity
}
