% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{lines,activity-method}
\alias{lines,activity-method}
\alias{lines,activity,activity-method}
\title{plot a Strava activtiy on OpenStreetmap}
\usage{
\S4method{lines}{activity}(x)
}
\arguments{
\item{x}{an object of S4 class activity.}
}
\description{
plot a Strava activtiy on OpenStreetmap
}
