% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lines.activity}
\alias{lines.activity}
\title{taking lat long coordinates of an object of S4 class activity and joining the corresponding points with line segments.}
\usage{
lines.activity(activity, ...)
}
\arguments{
\item{activity}{an object of S4 class activity.}
}
\description{
taking lat long coordinates of an object of S4 class activity and joining the corresponding points with line segments.
}
