% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lines_activity}
\alias{lines_activity}
\title{taking lat long coordinates of an object of S4 class activity and joining the corresponding points with line segments.}
\usage{
lines_activity(activity, ...)
}
\arguments{
\item{activity}{an object of S4 class activity.}

\item{...}{further arguments, see \code{\link{lines}}.}
}
\description{
taking lat long coordinates of an object of S4 class activity and joining the corresponding points with line segments.
}
