% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx.R
\name{load_activities}
\alias{load_activities}
\title{Load Strava activities from gpx files}
\usage{
load_activities(path = "activities/", type = "all", merge = FALSE,
  ...)
}
\arguments{
\item{path}{the path from de current working directory to a directory containing the gpx files from Strava.}

\item{type}{to restrict to a certain type of activities, like \code{"Ride"}, \code{"Run"}, etc.
By default \code{type = "all"}.}

\item{merge}{if TRUE, only an object of S4 class activity is returned, otherwise a list of activity object.
By default \code{merge = FALSE}.}

\item{...}{further arguments passed to the function \code{merge_activity} (only considered if \code{merge = TRUE}).}
}
\value{
either a list or a single activity object.
}
\description{
Load Strava activities from gpx files
}
