% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{makeQuery}
\alias{makeQuery}
\title{Make Query Call}
\usage{
makeQuery(query, token)
}
\arguments{
\item{query}{An appropriate API call, without access token for Strava API}

\item{token}{A character string of the user's access token}
}
\value{
result An httr response object from the Strava API call
}
\description{
Takes a query, attaches the access token, checks for errors, checks for rate
limits and returns the response object.
}

