% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_activity.R
\name{merge_activity}
\alias{merge_activity}
\title{Merge different objects of S4 class activtiy into one.}
\usage{
merge_activity(list_activity, name = "merged-activity")
}
\arguments{
\item{list_activity}{a list of activty objects.}

\item{name}{the name of the merged activity.
By default \code{name = "merged-activity"}.}
}
\value{
an object of S4 class activtiy.
}
\description{
Merge different objects of S4 class activtiy into one.
}
