% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{parse_gpx}
\alias{parse_gpx}
\title{Read and parse .gpx files from Strava.}
\usage{
parse_gpx(file)
}
\arguments{
\item{file}{a .gpx file from Strava.}
}
\value{
an object of S4 class activity, containing the slots: name, type, lat, lon, distances, elevations, times.
}
\description{
An object of S4 class activity is created.
Thereby, it is assumed, that the corresponding gpx file consists of one Strava activity.
}
