% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{parseActivityStream}
\alias{parseActivityStream}
\title{Parse Activity Stream}
\usage{
parseActivityStream(response)
}
\arguments{
\item{response}{An httr response object for a stream, unparsed}
}
\value{
data A dataframe object with all available fields
}
\description{
Parses an activity stream object and returns it as a data frame.
}

