% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{parseAthleteActivity}
\alias{parseAthleteActivity}
\title{Parse Athlete Activity}
\usage{
parseAthleteActivity(activityObj)
}
\arguments{
\item{activityObj}{A named list activity object from Strava JSON API}
}
\value{
out A data frame with one row with attributes from Strava API JSON
}
\description{
Parses through activity objects and returns as a single line data frame.
}

