% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{parseAthleteActivityList}
\alias{parseAthleteActivityList}
\title{Parse Athlete Activity list}
\usage{
parseAthleteActivityList(response)
}
\arguments{
\item{response}{An httr response object from an authenticated call to
Strava JSON API}
}
\value{
out A data frame with one row per activity with features from json
   object
}
\description{
Iterates through an Athlete Activity List and passes to
\link{parseAthleteActivity}.
}

