% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{parseAthleteShoeSummary}
\alias{parseAthleteShoeSummary}
\title{Parse Athlete's Shoe Summary Data}
\usage{
parseAthleteShoeSummary(response)
}
\arguments{
\item{response}{An httr response object from an authenticated call to
Strava JSON API}
}
\value{
out A data frame with one row per shoe object with features from json
   object
}
\description{
Parses Athlete's Shoe Summary data, drawn from Athlete JSON object from Strava
JSON API.
}
\examples{
##Not a real token
\dontrun{me <- retrieveAthlete('abcdefghijk')}
\dontrun{my_shoes <- parseAthleteShoeSummary(me)}
}

