% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,activity,ANY-method}
\alias{plot,activity,ANY-method}
\title{plot a Strava activtiy on OpenStreetmap}
\usage{
\S4method{plot}{activity,ANY}(x, y, ul, lr, osm_type, padby)
}
\arguments{
\item{x}{an object of S4 class activity.}

\item{y}{ANY.}

\item{ul}{upperleft lat and long.}

\item{lr}{lowerright lat and long.}

\item{osm_type}{one of the slot names, "name", "type", "lat", "lon", "distances", "elevations", "times".}

\item{padby}{pad upperleft and lowerright lat long points by this value.}

\item{...}{additional arguments for \code{plot}.}
}
\description{
plot a Strava activtiy on OpenStreetmap
}
