% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,activity,missing-method}
\alias{plot,activity,missing-method}
\alias{plot,activity,activity-method}
\title{plot a Strava activtiy on OpenStreetmap}
\usage{
\S4method{plot}{activity,missing}(x, y, ul, lr, osm_type, padby)
}
\arguments{
\item{x}{an object of S4 class activity.}

\item{y}{missing.}

\item{ul}{upperleft lat and long.}

\item{lr}{lowerright lat and long.}

\item{osm_type}{openstreetmap type, see \code{\link{openmap}} for possible values.}

\item{padby}{pad upperleft and lowerright lat long points by this value.}

\item{...}{additional arguments for \code{plot}.}
}
\description{
plot a Strava activtiy on OpenStreetmap
}
