% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.activity}
\alias{plot.activity}
\title{plot an activity, using OpenStreetMap.}
\usage{
plot.activity(activity, ul, lr, osm_type = "stamen-watercolor",
  padby = 0.01, ...)
}
\arguments{
\item{activity}{an object of S4 class activity.}

\item{ul}{upperleft lat and long.}

\item{lr}{lowerright lat and long.}

\item{osm_type}{openstreetmap type, see \code{\link{openmap}} for possible values.}

\item{padby}{pad upperleft and lowerright lat long points by this value.}

\item{...}{additional arguments for \code{plot}.}
}
\description{
plot an activity, using OpenStreetMap.
}
