% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_activity}
\alias{plot_activity}
\title{plot an activity, using OpenStreetMap.}
\usage{
plot_activity(activity, osm_type = "stamen-watercolor", padby = 0.01)
}
\arguments{
\item{activity}{an object of S4 class activity.}

\item{osm_type}{openstreetmap type, see \code{\link{openmap}}.}

\item{padby}{pad upperleft and lowerright latitude/longuitude points by this value.}
}
\description{
plot an activity, using OpenStreetMap.
}
