% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_listactivities}
\alias{plot_listactivities}
\title{plot a list of Strava Strava activities}
\usage{
plot_listactivities(lactivity, ul, lr, osm_type = "stamen-watercolor",
  padby = 0.01, ...)
}
\arguments{
\item{lactivity}{a list of activity objects.}

\item{ul}{upperleft lat and long.}

\item{lr}{lowerright lat and long.}

\item{osm_type}{one of the slot names, "name", "type", "lat", "lon", "distances", "elevations", "times".}

\item{padby}{pad upperleft and lowerright lat long points by this value.}
}
\description{
plot a list of Strava Strava activities
}
