% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{privatearea}
\alias{privatearea}
\title{add a circle to provide some privacy}
\usage{
privatearea(lat, lon, jitter = 1e-05, inch = 1/10, col = "white")
}
\arguments{
\item{lat}{a vector of lattitude coordinates.}

\item{lon}{a vector of longuitude coordinates.}

\item{jitter}{a rondom jitter, such that the center of the circle is not at the exact location to protect.}

\item{inch}{determines the circle size.}

\item{col}{circle color.}
}
\description{
add a circle to provide some privacy
}
