% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{readToken}
\alias{readToken}
\title{Read User Access Token From File}
\usage{
readToken(filename)
}
\arguments{
\item{filename}{String containing path to file containing plain text token}
}
\value{
token String value for the token
}
\description{
Read User Access Token From File
}
\examples{
\dontrun{token <- readToken(/path/to/file.txt)}
}

