% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{retrieveActivityStream}
\alias{retrieveActivityStream}
\title{Retrieve Activity Stream}
\usage{
retrieveActivityStream(token, activity_id)
}
\arguments{
\item{token}{String value for user access token to API}

\item{activity_id}{(optional) Athlete ID, if retreiving non-authenticated user}
}
\value{
result An httr response object, unparsed
}
\description{
Takes an activity id and returns the stream (raw data) of the activity
}

