% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{retrieveAllActivityStreamsIntoDf}
\alias{retrieveAllActivityStreamsIntoDf}
\title{Retrieve All Activity Streams Into Data Frame}
\usage{
retrieveAllActivityStreamsIntoDf(token, activity_df)
}
\arguments{
\item{token}{String value for user access token to API}

\item{activity_df}{A dataframe object with all available fields}
}
\value{
data A dataframe object with all available fields, stacking by
 activity id
}
\description{
Reads through an activity data frame, making a request to API for each
activity id, and combines it into a tidy data frame with a unique record
for each observation and activity id.
}
\examples{
token <- readToken('path/to/token.txt')
activity_response <- retrieveAthleteActivities(token)
activity_df <- parseAthleteActivities(activity_response)
activity_stream_df <- retrieveAllActivityStreamsIntoDf(token, activity_df)
}

