% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{retrieveAthlete}
\alias{retrieveAthlete}
\title{Retrieve Athlete Information from API}
\usage{
retrieveAthlete(token, athlete_id = NULL)
}
\arguments{
\item{token}{String value for user access token to API}

\item{athlete_id}{(optional) Athlete ID, if retreiving non-authenticated user}
}
\value{
result An httr response object, unparsed
}
\description{
\code{retrieveAthlete} fetches athlete information from the Strava API, checks
for errors and the current rate limit, and return the result (full response).
Will raise warnings if the response status is not 200, and also if the user
has used 90% of their API query limit (short term or long term).
}
\examples{
##Not a real token
\dontrun{me <- retrieveAthlete('abcdefghijk')}
\dontrun{someoneElse <- retrieveAthlete('abcdefghijk', athlete_id='1234')}
}

