% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{retrieveAthleteActivities}
\alias{retrieveAthleteActivities}
\title{Retrieve Athlete Activities}
\usage{
retrieveAthleteActivities(token, before = NULL, after = NULL)
}
\arguments{
\item{token}{String value for user access token to API}

\item{before}{(optional) Integer value representing seconds since epoch}

\item{after}{(optional) Integer value representing seconds since epoch}
}
\value{
result An httr response object, unparsed
}
\description{
Retrieves athlete activities, in one hit. If it will be more than 300
activities, the function will need to be modified.
}

