% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{retrieveAthleteActivitySpecific}
\alias{retrieveAthleteActivitySpecific}
\title{Retrieve Specific Athlete Activity}
\usage{
retrieveAthleteActivitySpecific(token, activity_id)
}
\arguments{
\item{token}{String value for user access token to API}

\item{activity_id}{Unique identifier for the activity, as an integer or string}
}
\value{
An unparsed HTTR response object
}
\description{
Using activity_id, downloads the detailed resource_state for an activity
}

