% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{retrieveAthleteStats}
\alias{retrieveAthleteStats}
\title{Retrieves Athlete Stats}
\usage{
retrieveAthleteStats(token, athlete_id)
}
\arguments{
\item{token}{String value for user access token to API}

\item{athlete_id}{Athlete ID for user of interest}
}
\value{
result An httr response object, unparsed
}
\description{
Retrieves Athlete Stats
}

