% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava_auth.R
\name{strava_auth}
\alias{strava_auth}
\title{Generata Strava API authentication token}
\usage{
strava_auth(app_name = Sys.getenv("STRAVA_APP_NAME"),
  app_client_id = Sys.getenv("STRAVA_APP_ID"),
  app_secret = Sys.getenv("STRAVA_APP_SECRET"), token = NULL,
  new_user = FALSE)
}
\arguments{
\item{app_name}{Name of your developer app}

\item{app_client_id}{The Client ID of your developer app}

\item{app_secret}{The secret for your developer app}
}
\description{
Generate a token for the user and the desired scope. The user is sent to the strava authentication page if he/she hasn't given permission to the app yet, else, is sent to the app webpage.
}
