% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava_customApi.R
\name{strava_customApi}
\alias{strava_customApi}
\title{Get custom API request string}
\usage{
strava_customApi(request)
}
\arguments{
\item{request}{Request string after https://www.strava.com/api/v3/, e.g. athletes}
}
\description{
Type in your own request string
}
\examples{
strava_customApi(request)
}
