% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava_laps.R
\name{strava_laps}
\alias{strava_laps}
\title{Get laps from Strava}
\usage{
strava_laps(data, lapType = "custom")
}
\arguments{
\item{data}{The activity id or vector of ids}

\item{lapType}{Wether to import custom laps from data, or standard. Defaults to custom. c("custom", "standard")}
}
\description{
Get laps information by the specific id or vector of ids
}
\examples{
strava_laps(12345678, lapType = c("custom", "standard"))
}
