% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava_map.R
\name{strava_map}
\alias{strava_map}
\title{Produce leaflet map with Strava data}
\usage{
strava_map(long, lat, zoom)
}
\arguments{
\item{long}{Initial longitude position}

\item{lat}{Initial latitude position}

\item{zoom}{Initial zoom level}
}
\value{
a leaflet map
}
\description{
Produce leaflet map with Strava data
}
