\name{StravaR-package}
\alias{StravaR-package}
\alias{StravaR}
\docType{package}
\title{
\packageTitle{StravaR}
}
\description{
\packageDescription{StravaR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{StravaR}
\packageIndices{StravaR2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{StravaR}

Maintainer: \packageMaintainer{StravaR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
