% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\docType{methods}
\name{[,activity,ANY,ANY,ANY-method}
\alias{[,activity,ANY,ANY,ANY-method}
\title{Definition of the subset operator for an object of S4 class activity.}
\usage{
\S4method{[}{activity,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{an object of S4 class activity.}

\item{i}{one of the slot names, "name", "type", "lat", "lon", "distances", "elevations", "times".}
}
\value{
the respective slot entries.
}
\description{
Definition of the subset operator for an object of S4 class activity.
}
