% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StravaR.R
\name{updateActivitiesWithDetail}
\alias{updateActivitiesWithDetail}
\title{Update Activities with Detail}
\usage{
updateActivitiesWithDetail(token, activity_df)
}
\arguments{
\item{token}{String value for user access token to API}

\item{activity_df}{Dataframe containing parsed activities}
}
\value{
activity_df Dataframe with updated, detailed parsed activities
}
\description{
Uses activity summary data frame and then queries each indiidual activity
to get a detailed (and if changed, updated) record for that activity. It
replaces the old records and returns it as a dataframe.
}

