% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava_activities.R
\name{strava_activities}
\alias{strava_activities}
\title{Get activities from Strava}
\usage{
strava_activities()
}
\arguments{
\item{scope}{Name of your developer app}

\item{token}{The Client ID of your developer app}

\item{app_secret}{The secret for your developer app}
}
\description{
Get activities from the authenticated user as a data frame
}
